requirements.txt
# Dependências principais do projeto
evolutionapi==0.0.5
groq==0.18.0
openai==1.66.3
unidecode==1.3.8

# Dependências de API e requisições
requests==2.32.3
requests-toolbelt==1.0.0
requests-oauthlib==2.0.0
httpx==0.28.1
httpcore==1.0.9

# Dependências de IA e processamento de texto
langchain-core==0.3.62
langchain-text-splitters==0.3.8
langsmith==0.3.42
tokenizers==0.21.1
huggingface-hub==0.32.2
rapidfuzz==3.8.1

# Dependências de FastAPI e servidor
fastapi==0.115.9
fastapi-cli==0.0.7
uvicorn==0.34.2
starlette==0.45.3
asgiref==3.8.1

# Dependências de banco de dados
psycopg2-binary==2.9.10
chromadb-client==1.0.12

# Dependências de processamento de documentos
PyMuPDF==1.26.0
python-multipart==0.0.20

# Dependências de autenticação e segurança
passlib==1.7.4
bcrypt==4.0.1
PyJWT==2.10.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0

# Dependências de utilitários
python-dotenv==1.1.0
pydantic==2.11.5
pydantic_core==2.33.2
python-dateutil==2.9.0.post0

# Dependências de monitoramento e telemetria
opentelemetry-api==1.33.1
opentelemetry-sdk==1.33.1
opentelemetry-instrumentation-fastapi==0.54b1

# Dependências de processamento de texto e números
num2words==0.5.12

# Dependências de armazenamento
minio==7.2.15

# Dependências de sistema e utilitários
anyio==4.9.0
attrs==25.3.0
backoff==2.2.1
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
coloredlogs==15.0.1
Deprecated==1.2.18
distro==1.9.0
dnspython==2.7.0
durationpy==0.10
email_validator==2.2.0
filelock==3.18.0
flatbuffers==25.2.10
fsspec==2025.5.1
google-auth==2.40.2
googleapis-common-protos==1.70.0
grpcio==1.71.0
h11==0.16.0
hf-xet==1.1.2
httptools==0.6.4
humanfriendly==10.0
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kubernetes==32.0.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mmh3==5.1.0
mpmath==1.3.0
numpy==2.2.6
oauthlib==3.2.2
onnxruntime==1.22.0
opentelemetry-exporter-otlp-proto-common==1.33.1
opentelemetry-exporter-otlp-proto-grpc==1.33.1
opentelemetry-instrumentation==0.54b1
opentelemetry-instrumentation-asgi==0.54b1
opentelemetry-proto==1.33.1
opentelemetry-semantic-conventions==0.54b1
opentelemetry-util-http==0.54b1
orjson==3.10.18
overrides==7.7.0
packaging==24.2
posthog==4.2.0
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.23.0
Pygments==2.19.1
PyPika==0.48.9
pyproject_hooks==1.2.0
PyYAML==6.0.2
referencing==0.36.2
rich==14.0.0
rich-toolkit==0.14.7
rpds-py==0.25.1
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
tenacity==9.1.2
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
uvloop==0.21.0
watchfiles==1.0.5
websocket-client==1.8.0
websockets==15.0.1
wrapt==1.17.2
zipp==3.22.0
zstandard==0.23.0